requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
streamlit
pytest
langgraph
langchain
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
langchain-huggingface
langchain-pinecone
pinecone
